add no-sample error fix

+++ bu_isciii/clean.py
@@ -65,7 +65,6 @@ class CleanUp:
         self.services_requested = self.resolution_info["Resolutions"][
             "availableServices"
         ]
-        self.service_samples = self.resolution_info["Samples"]
         if self.service_folder in self.path:
             self.full_path = self.path
         else:
@@ -81,7 +80,7 @@ class CleanUp:
         self.delete_files = self.get_clean_items(self.services_to_clean, type="files")
         # self.delete_list = [item for item in self.delete_list if item]
         self.nocopy = self.get_clean_items(self.services_to_clean, type="no_copy")
-        self.service_samples = self.resolution_info["Samples"]
+        self.service_samples = self.resolution_info.get("Samples", None)
 
         if option is None:
             self.option = bu_isciii.utils.prompt_selection(
@@ -302,15 +301,16 @@ class CleanUp:
         Params:
 
         """
-        files_to_delete = []
-        for sample_info in self.service_samples:
-            for file in self.delete_files:
-                file_to_delete = file.replace("sample_name", sample_info["sampleName"])
-                files_to_delete.append(file_to_delete)
-        path_content = self.scan_dirs(to_find=files_to_delete)
-        for file in path_content:
-            os.remove(file)
-            stderr.print("[green]Successfully removed " + file)
+        if self.service_samples is not None:
+            files_to_delete = []
+            for sample_info in self.service_samples:
+                for file in self.delete_files:
+                    file_to_delete = file.replace("sample_name", sample_info["sampleName"])
+                    files_to_delete.append(file_to_delete)
+            path_content = self.scan_dirs(to_find=files_to_delete)
+            for file in path_content:
+                os.remove(file)
+                stderr.print("[green]Successfully removed " + file)
         return
 
     def purge_folders(self, sacredtexts=["lablog", "logs"], add="", verbose=True):
